dom/simplepush/test/test_permissions.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1235050: SimplePush permissions tests.

Any copyright is dedicated to the Public Domain.
http://creativecommons.org/licenses/publicdomain/

-->
<head>
  <title>Test for Bug 1235050</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta http-equiv="Content-type" content="text/html;charset=UTF-8">
</head>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1235050">Mozilla Bug 1235050</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>

<script class="testbody" type="text/javascript">


  function already_has_permission() {
    return SpecialPowers.hasPermission("push", document);
  }

  function add_permission() {
    var allow = SpecialPowers.Ci.nsIPermissionManager.ALLOW_ACTION;
    SpecialPowers.addPermission("push", allow, document);
    window.location.reload();
  }

  function test_no_permission() {
    ok((!("push" in navigator)), "push object is NOT exposed");
  }

  function test_has_permission() {
    ok(("push" in navigator), "push object is exposed");
  }

  function runTest() {
    if (already_has_permission()) {
      test_has_permission();
      SimpleTest.finish();
    } else {
      test_no_permission();
      add_permission();
    }
  }

  SimpleTest.waitForExplicitFinish();

  runTest();
</script>
</body>
</html>
